Commit the dependencies to make a new git checkout easier to use
This is probably inadvisible for most projects, but some guys
expressed a desire to use the git version of Akku and wanted it to be
easier. A fresh checkout should now work (with Chez) without any other
preparation.

The files added here are not *everything* under .akku/lib, but only
what akku dependency-scan said was needed.