Library name mangling, omitting and blocking; handle SRFIs for Guile
This adds code in schemedb to mangle library names during installation
and omit/block installation of libraries. These features are
per-implementation and can work around quirks in implementations.

SRFI-97 based library names are mangled to match Guile's own mangling
for these libraries. Some SRFIs are blocked for Guile because they are
needed during startup of the runtime and can't be provided
non-natively.