Do not update timestamps on unchanged files during installation
Some implementations like Guile check the timestamp on files to
determine if they need to be recompiled. The installer now preserves
timestamps on files that haven't changed by writing to a temporary
file, then comparing with the previous version of the file and only
renaming the temporary file if the contents differ.