Fix for updating a shallow clone
When a repository was installed using a shallow clone (i.e. when there
was a git tag to pull) and the project later changed to not having a
tag (after a new lock), the existing clone was unable to fetch the new
revision.

A workaround is to do rm -rf .akku/src/project-name.