Reworked scripts support
After some more experimentation the scripts are now run only once per
occasion (no longer once per occasion + mentioned implementation). The
scripts also have access to a few environment variables for the paths
to libraries, etc. This should be flexible enough to do anything at
all, basically, but more importantly to have something which you can
experiment with.

The operating system name is available to the cond-expand of scripts
(as the symbol linux, cygwin, etc, the lower-casing of uname -s). The
idea isn't to make a complete build system, but just to provide enough
info to pick the right set of scripts.