Support for direct dependencies in the manifest
This adds preliminary support for direct dependencies in the manifest.
These can be useful during development to use dependencies that are
not yet in the package index.

Instead of writing a SemVer range for the dependency, you need to
write out an explicit version, its location and revision (and
optionally a tag). Example usage in Akku.manifest to get an unreleased
version of "text-mode":

  (akku-package ...
    ...
    (depends ...
             ("text-mode"
              (version "0.0.0")
              (location (git "https://gitlab.com/weinholt/text-mode"))
              (revision "bf7dac89135a5cb8379c82ccd5d2c86b6db33a66"))))

Documentation and UI remain to be done.

Fixes: #8