Refresh dependencies to fix tar unpacking; accept '..' in tarfiles
This is a refresh of the dependencies that mainly fix a bug in the
(compression tar) library. Akku's tar unpacking is also changed to
handle '..' in filenames, still relying on check-filename to make sure
that files do not go outside of the repository (although all files are
also vetted before going into the package index).