Recognize git ssh url's without ssh://, various fixes
Fixes support for git@github:etc URLs (default in GitHub's UI). Fixes
a crash when location was missing from the manifest. Fixes the
location override not being used in untagged releases. Fixes an ugly
warning when there was no recognized remote. Fixes a weird thing that
happened if no commits were on the tree.