Order the lockfile according to the manifest
The lockfile is now ordered in a way that lets the manifest control
the installation order. This is useful because projects which are
installed later can override earlier projects. Unmentioned
dependencies are installed first, then the dependencies from the
manifest (in the same order as the manifest) and lastly the current
project.