Add files for the GNU build system
This adds (optional) support for building with autotools and GNU
Guile. This makes it easy to do a system-wide installation.

It should also make it easier for distributors to package Akku, making
this relevant to issues #18, #20 and #22.

The setup is based on this article:
https://www.erikedrosa.com/2017/10/29/guile-projects-with-autotools.html

This commit also simplifies the license info to make it easier for
downstreams.

And, and, and! This commit fixes the dependency-scan subcommand, so
that it handles multiple implementations correctly and supports the
library exclusion lists.

Fixes: #21