Copy comments when copying/translating code from libraries
The translated R7RS libraries previously lost all their comments. They
now have the comments from the top of the file. This is mostly done to
preserve copyright notices.

R6RS libraries that are extracted from a file with multiple libraries
are now preserved with comments, including those from the top of the
file.